+++ src/features/share-schedule/index.ts
@@ -1,0 +1,1 @@
+export { ShareSchedule } from "./ui";
