+++ src/features/share-schedule/model/types.ts
@@ -1,0 +1,7 @@
+export interface ShareScheduleItemType {
+  id: number;
+  title: string;
+  date: string;
+  startTime: string;
+  endTime: string;
+}
