+++ src/features/view-daily-schedule/index.ts
@@ -1,0 +1,1 @@
+export { DailySchedule } from "./ui";
