+++ src/features/view-daily-schedule/model/types.ts
@@ -1,0 +1,6 @@
+import type { ScheduleCalendar } from "@/entities/calendar";
+import type { Schedule } from "@/entities/schedule";
+
+export type ScheduleItem = Pick<Schedule, "id" | "title" | "location" | "startTime" | "endTime" | "isAllDay"> & {
+  calendar: ScheduleCalendar;
+};
