+++ src/shared/model/types.ts
@@ -1,0 +1,1 @@
+export type RightSidebarViewType = "daily" | "share" | "create" | "edit" | "search" | "notification";
