--- src/widgets/RightSidebar/ui/index.ts
@@ -1,2 +1,0 @@
-export { ToggleCheckButton } from "./ToggleCheckButton";
-export { ShareScheduleItem } from "./ShareScheduleItem";
